# qtai-server/apis/api-v1/openapi.yaml
openapi: 3.0.3
info:
  title: QT-AI API
  version: "1.0.0"
  description: QT-AI 서버 v1 OpenAPI 명세 (각 도메인 PR에서 점진적으로 채워나감)
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/admin/ai/prompt-versions:
    get:
      tags:
        - Admin AI
      operationId: listAdminAiPromptVersions
      summary: List admin AI prompt versions
      description: >
        Lists AI prompt versions. The first managed scope supports EXPLANATION only.
        Requires ROLE_ADMIN and one of ADMIN_ROLE_REVIEWER or ADMIN_ROLE_SUPER_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: promptType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AiPromptType"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AiPromptVersionStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Admin AI prompt version list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAiPromptVersionListApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Admin AI
      operationId: createAdminAiPromptVersion
      summary: Create admin AI prompt version
      description: >
        Creates a DRAFT EXPLANATION prompt version. The prompt can become ACTIVE
        only after a successful evaluation run with no failed or needs-review cases.
        Requires ROLE_ADMIN and one of ADMIN_ROLE_REVIEWER or ADMIN_ROLE_SUPER_ADMIN.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAiPromptVersionRequest"
      responses:
        "201":
          description: Admin AI prompt version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAiPromptVersionApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/admin/ai/prompt-versions/{id}:
    get:
      tags:
        - Admin AI
      operationId: getAdminAiPromptVersion
      summary: Get admin AI prompt version
      description: >
        Returns one AI prompt version including prompt text and runtime settings.
        Requires ROLE_ADMIN and one of ADMIN_ROLE_REVIEWER or ADMIN_ROLE_SUPER_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Admin AI prompt version returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAiPromptVersionApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/admin/ai/prompt-versions/{id}/activate:
    post:
      tags:
        - Admin AI
      operationId: activateAdminAiPromptVersion
      summary: Activate admin AI prompt version
      description: >
        Activates a DRAFT EXPLANATION prompt version after a successful evaluation
        run. Existing ACTIVE prompt versions of the same type are retired.
        Requires ROLE_ADMIN and one of ADMIN_ROLE_REVIEWER or ADMIN_ROLE_SUPER_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Admin AI prompt version activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAiPromptVersionApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/admin/ai/prompt-versions/{id}/retire:
    post:
      tags:
        - Admin AI
      operationId: retireAdminAiPromptVersion
      summary: Retire admin AI prompt version
      description: >
        Retires an EXPLANATION prompt version. Requires ROLE_ADMIN and one of
        ADMIN_ROLE_REVIEWER or ADMIN_ROLE_SUPER_ADMIN.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Admin AI prompt version retired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAiPromptVersionApiResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        관리자 AI API는 ROLE_ADMIN과 세부 ADMIN_ROLE authority를 확인한다. 시스템 API는
        SYSTEM_BATCH 또는 ROLE_SYSTEM_BATCH authority를 확인하며, service account 기반 검증
        필터는 아직 구현 범위 밖이다.
...
  responses:
    BadRequest:
      description: 요청 값이 올바르지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    Unauthorized:
      description: 인증이 필요함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    Forbidden:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    NotFound:
      description: 요청한 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    Conflict:
      description: 상태 전이 또는 중복 처리 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorApiResponse"
  schemas:
...
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorApiResponse:
      type: object
      required:
        - success
        - data
        - error
        - timestamp
        - traceId
      properties:
        success:
          type: boolean
          enum:
            - false
        data:
          type: object
          nullable: true
        error:
          $ref: "#/components/schemas/ErrorBody"
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
...
    AdminAiPromptVersionListApiResponse:
      type: object
      required:
        - success
        - data
        - error
        - timestamp
        - traceId
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AdminAiPromptVersionListResponse"
        error:
          allOf:
            - $ref: "#/components/schemas/ErrorBody"
          nullable: true
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
    AdminAiPromptVersionApiResponse:
      type: object
      required:
        - success
        - data
        - error
        - timestamp
        - traceId
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AdminAiPromptVersionResponse"
        error:
          allOf:
            - $ref: "#/components/schemas/ErrorBody"
          nullable: true
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
...
    AdminAiPromptVersionRequest:
      type: object
      required:
        - promptType
        - version
        - systemPrompt
        - userPromptTemplate
      properties:
        promptType:
          $ref: "#/components/schemas/AiPromptType"
        version:
          type: string
          minLength: 1
          maxLength: 30
          example: "2026.06.2"
        systemPrompt:
          type: string
          minLength: 1
          description: System prompt used for LLM completion.
        userPromptTemplate:
          type: string
          minLength: 1
          description: >
            User prompt template. Supported variables include {{targetType}},
            {{targetId}}, {{qtPassageBlock}}, {{versesBlock}}, and {{commentaryBlock}}.
        modelName:
          type: string
          maxLength: 100
          nullable: true
          example: deepseek-chat
        temperature:
          type: number
          format: double
          minimum: 0
          maximum: 2
          nullable: true
          default: 0.2
        maxTokens:
          type: integer
          format: int32
          minimum: 1
          maximum: 20000
          nullable: true
          default: 2000
        description:
          type: string
          maxLength: 500
          nullable: true
    AdminAiPromptVersionListResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
        - first
        - last
        - sort
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/AdminAiPromptVersionResponse"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
        first:
          type: boolean
        last:
          type: boolean
        sort:
          type: string
          enum:
            - createdAt,desc,id,desc
    AdminAiPromptVersionResponse:
      type: object
      required:
        - id
        - promptType
        - version
        - contentHash
        - status
        - systemPrompt
        - userPromptTemplate
        - modelName
        - temperature
        - maxTokens
        - description
        - createdByAdminId
        - createdAt
        - activatedAt
        - retiredAt
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        promptType:
          $ref: "#/components/schemas/AiPromptType"
        version:
          type: string
          maxLength: 30
        contentHash:
          type: string
          maxLength: 100
        status:
          $ref: "#/components/schemas/AiPromptVersionStatus"
        systemPrompt:
          type: string
        userPromptTemplate:
          type: string
        modelName:
          type: string
          nullable: true
        temperature:
          type: number
          format: double
        maxTokens:
          type: integer
          format: int32
        description:
          type: string
          nullable: true
        createdByAdminId:
          type: integer
          format: int64
          nullable: true
        createdAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
          nullable: true
        retiredAt:
          type: string
          format: date-time
          nullable: true
...
    AiPromptType:
      type: string
      enum:
        - EXPLANATION
        - SIMULATOR
        - QA
    AiPromptVersionStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - RETIRED
...